requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
PyPDF2==3.0.1
pytz==2023.3
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
Unidecode==1.3.6
urllib3==1.26.14
